move DEFAULT_MODELS + DEFAULT_PREDICTORS into commands/serve and commands/predict
