+++ allennlp/commands/serve.py
@@ -28,9 +28,36 @@ from allennlp.commands.subcommand import Subcommand
 from allennlp.service import server_sanic
 from allennlp.service.predictors import DemoModel
 
+# This maps from the name of the task
+# to the ``DemoModel`` indicating the location of the trained model
+# and the type of the ``Predictor``.  This is necessary, as you might
+# have multiple models (for example, a NER tagger and a POS tagger)
+# that have the same ``Predictor`` wrapper. The corresponding model
+# will be served at the `/predict/<name-of-task>` API endpoint.
+DEFAULT_MODELS = {
+        'machine-comprehension': DemoModel(
+                'https://s3-us-west-2.amazonaws.com/allennlp/models/bidaf-model-2017.09.15-charpad.tar.gz',  # pylint: disable=line-too-long
+                'machine-comprehension'
+        ),
+        'semantic-role-labeling': DemoModel(
+                'https://s3-us-west-2.amazonaws.com/allennlp/models/srl-model-2017.09.05.tar.gz', # pylint: disable=line-too-long
+                'semantic-role-labeling'
+        ),
+        'textual-entailment': DemoModel(
+                'https://s3-us-west-2.amazonaws.com/allennlp/models/decomposable-attention-2017.09.04.tar.gz',  # pylint: disable=line-too-long
+                'textual-entailment'
+        ),
+        'coreference-resolution': DemoModel(
+                'https://s3-us-west-2.amazonaws.com/allennlp/models/coref-model-2017.11.09.tar.gz',  # pylint: disable=line-too-long
+                'coreference-resolution'
+        )
+}
+
+
 class Serve(Subcommand):
-    def __init__(self, trained_models: Dict[str, DemoModel]) -> None:
-        self.trained_models = trained_models
+    def __init__(self, model_overrides: Dict[str, DemoModel] = {}) -> None:
+        # pylint: disable=dangerous-default-value
+        self.trained_models = {**DEFAULT_MODELS, **model_overrides}
 
     def add_subparser(self, name: str, parser: argparse._SubParsersAction) -> argparse.ArgumentParser:
         # pylint: disable=protected-access
